Fix bug where Fxx speed effects slow down the song by one tick
Partly fixes Fxx tempo effect slowdown as well.

TickCounter::countUp() calls TickCounter::resetRest() to compute the
length of the next row and write it to restTickToNextStep_. However,
when calling resetRest() during song playback, restTickToNextStep_ must
be decremented by 1 afterwards so the row runs for the correct number of
ticks (since TickCounter::countUp() only runs the next row when called
with restTickToNextStep_ set to 0, otherwise it decrements it).

In the existing code, TickCounter::countUp() first calls resetRest(),
then decrements restTickToNextStep_. However, when processing a Fxx
effect, after TickCounter::countUp() is called,
TickCounter::setSpeed(int speed) gets called, which calls resetRest()
*without* decrementing restTickToNextStep_.

I changed TickCounter::resetRest() to decrement restTickToNextStep_ if
the song is playing. I'm not sure if this is the "right" approach, but
it fixes this bug and doesn't have any negative effects as far as I can
tell.

There are several cases when resetRest() can be called:

- When loading file (resetRest() must not decrement restTickToNextStep_)

- When starting playback through PlaybackManager::playStep() or
  PlaybackManager::startPlay() (doesn't matter, they call resetCount()
  which sets restTickToNextStep_ to 0)

- While starting playback, when retrieving speed/tempo/groove effects
  (restTickToNextStep_ doesn't matter, will be overwritten by countUp())

- After countUp() processes head of step (resetRest() must decrement
  restTickToNextStep_)

- When changing a song's global speed, tempo, or groove, including when
  playback is active (the exact behavior is unimportant, my actual
  concern is that this currently data-races)

This commit makes resetRest() decrement restTickToNextStep_ if the song
is playing.

There are other cases when resetRest() is never called (I think) unless
you change global settings:

- Before countUp() processes head of step (the current code always calls
  countUp() before calling resetRest())

- After countUp() processes middle of step (never happens, Fxx effects
  are always processed at the head, not delayed by a Gxx effect)